requirements.txt
ircchallenge ~=0.1.2
ircrobots    ~=0.6.2
pyyaml       ~=5.4.1

